python-services/payment-worker/requirements.txt
apscheduler>=3.10.0
psycopg2-binary>=2.9.9
pandas>=2.2.0
kafka-python>=2.0.0
